Requisição para adicionar uma música à playlist selecionada, funcionando
