+++ src/components/GenreChipGroup/GenreChipGroup.js
@@ -1,8 +1,5 @@
 import React from 'react';
 import { makeStyles } from '@material-ui/core/styles';
-import Chip from '@material-ui/core/Chip';
-import Paper from '@material-ui/core/Paper';
-import TagFacesIcon from '@material-ui/icons/TagFaces';
 import { ChipGroupContainer, StyledChip } from './styles';
 
 const useStyles = makeStyles((theme) => ({
